Fix bug "1" ≠ 1 Fix bug countdown interval

+++ app/Http/Controllers/PhotoController.php
@@ -25,7 +25,7 @@ class PhotoController extends Controller
             'triggerUrl' => $triggerUrl,
             'bearerToken' => $request->user()->createToken('DSLR')->plainTextToken,
             'uploadUrl' => route('photos.store'),
-            'timerSeconds' => nova_get_setting('timer', 3),
+            'timerSeconds' => (int)nova_get_setting('timer', 3),
         ]);
     }
